Move query string parsing into core
