+++ src/url.js
@@ -19,10 +19,11 @@ import {dict, hasOwn} from './core/types/object';
 import {endsWith} from './core/types/string';
 import {getMode} from './mode';
 import {isArray} from './core/types';
-import {parseQueryString_} from './url-parse-query-string';
-import {tryDecodeUriComponent_} from './url-try-decode-uri-component';
+import {parseQueryString, tryDecodeUriComponent} from './core/types/string/url';
 import {urls} from './config';
 import {userAssert} from './log';
+
+export {parseQueryString, tryDecodeUriComponent};
 
 /**
  * @type {!JsonObject}
@@ -353,20 +354,6 @@ export function assertAbsoluteHttpOrHttpsUrl(urlString) {
 }
 
 /**
- * Parses the query string of an URL. This method returns a simple key/value
- * map. If there are duplicate keys the latest value is returned.
- *
- * This function is implemented in a separate file to avoid a circular
- * dependency.
- *
- * @param {string} queryString
- * @return {!JsonObject}
- */
-export function parseQueryString(queryString) {
-  return parseQueryString_(queryString);
-}
-
-/**
  * Returns the URL without fragment. If URL doesn't contain fragment, the same
  * string is returned.
  * @param {string} url
@@ -652,18 +639,6 @@ export function checkCorsUrl(url) {
     'Source origin is not allowed in %s',
     url
   );
-}
-
-/**
- * Tries to decode a URI component, falling back to opt_fallback (or an empty
- * string)
- *
- * @param {string} component
- * @param {string=} opt_fallback
- * @return {string}
- */
-export function tryDecodeUriComponent(component, opt_fallback) {
-  return tryDecodeUriComponent_(component, opt_fallback);
 }
 
 /**
